Add vision pipeline and major refactor
