+++ mistralrs-core/src/pipeline/macros.rs
@@ -231,6 +231,28 @@ macro_rules! normal_model_loader {
 }
 
 #[macro_export]
+macro_rules! vision_normal_model_loader {
+    ($paths:expr, $dtype:expr, $default_dtype:expr, $device:expr, $config:expr, $loader:expr, $use_flash_attn:expr, $silent:expr, $mapper:expr, $loading_isq:expr, $real_device:expr) => {{
+        let vb = from_mmaped_safetensors(
+            $paths.get_weight_filenames().to_vec(),
+            Vec::new(),
+            $dtype.unwrap_or($default_dtype),
+            $device,
+            $silent,
+        )?;
+
+        $loader.load(
+            &$config,
+            $use_flash_attn,
+            vb,
+            $mapper,
+            $loading_isq,
+            $real_device,
+        )?
+    }};
+}
+
+#[macro_export]
 macro_rules! xlora_model_loader {
     ($paths:expr, $dtype:expr, $default_dtype:expr, $device:expr, $config:expr, $loader:expr, $use_flash_attn:expr, $silent:expr, $mapper:expr, $loading_isq:expr, $real_device:expr) => {{
         let mut safetensors_paths = $paths.get_weight_filenames().iter().collect::<Vec<_>>();
