+++ mistralrs-core/src/pipeline/speculative.rs
@@ -1,4 +1,5 @@
 use std::{
+    any::Any,
     iter::zip,
     sync::{Arc, Mutex},
 };
@@ -20,9 +21,10 @@ use crate::{
 };
 
 use super::{
-    cache_manager::DefaultCacheManager, calculate_inputs, chat_template::ChatTemplate,
-    sampling::SpeculativeSample, CacheInstruction, CacheManager, GeneralMetadata, ModelInputs,
-    ModelPaths,
+    cache_manager::DefaultCacheManager, chat_template::ChatTemplate,
+    inputs_processor::InputsProcessor, sampling::SpeculativeSample, AdapterActivationMixin,
+    CacheInstruction, CacheManager, CacheManagerMixin, GeneralMetadata, IsqPipelineMixin,
+    MetadataMixin, ModelPaths, PreProcessingMixin,
 };
 
 pub struct SpeculativeLoader {
@@ -153,7 +155,7 @@ impl SpeculativePipeline {
             candle_core::bail!("Target and draft models' tokenzier vocab do not match. This is required for speculative decoding.");
         }
         let metadata = get_mut_arcmutex!(target).get_metadata().clone();
-        // todo: some checks or relaxation here?
+        // TODO: some checks or relaxation here?
         Ok(Self {
             target,
             draft,
@@ -164,8 +166,101 @@ impl SpeculativePipeline {
     }
 }
 
+impl PreProcessingMixin for SpeculativePipeline {
+    fn get_chat_template(&self) -> Arc<ChatTemplate> {
+        get_mut_arcmutex!(self.target).get_chat_template()
+    }
+    fn get_input_processor(&self) -> &dyn InputsProcessor {
+        todo!()
+    }
+}
+
+impl IsqPipelineMixin for SpeculativePipeline {
+    fn re_isq_model(&mut self, dtype: GgmlDType) -> anyhow::Result<()> {
+        get_mut_arcmutex!(self.target).re_isq_model(dtype)?;
+        get_mut_arcmutex!(self.draft).re_isq_model(dtype)
+    }
+}
+
+impl CacheManagerMixin for SpeculativePipeline {
+    fn clone_in_cache(&mut self, seqs: &mut [&mut Sequence], modify_draft_cache: bool) {
+        DefaultCacheManager.clone_in_cache(
+            &mut *get_mut_arcmutex!(self.draft),
+            seqs,
+            modify_draft_cache,
+        );
+        DefaultCacheManager.clone_in_cache(&mut *get_mut_arcmutex!(self.target), seqs, false);
+    }
+    fn clone_out_cache(&mut self, seqs: &mut [&mut Sequence], modify_draft_cache: bool) {
+        DefaultCacheManager.clone_out_cache(
+            &mut *get_mut_arcmutex!(self.draft),
+            seqs,
+            modify_draft_cache,
+        );
+        DefaultCacheManager.clone_out_cache(&mut *get_mut_arcmutex!(self.target), seqs, false);
+    }
+    fn set_none_cache(&mut self, reset_non_granular: bool, modify_draft_cache: bool) {
+        DefaultCacheManager.set_none_cache(&mut *get_mut_arcmutex!(self.draft), modify_draft_cache);
+        DefaultCacheManager.set_none_cache(&mut *get_mut_arcmutex!(self.target), false);
+        if reset_non_granular {
+            self.reset_non_granular_state()
+        }
+        self.latest_logit_cache = None;
+    }
+    fn cache(&self) -> &Cache {
+        unreachable!()
+    }
+}
+
+impl AdapterActivationMixin for SpeculativePipeline {
+    /// Returns the number of activated adapters.
+    fn activate_adapters(&mut self, adapters: Vec<String>) -> anyhow::Result<usize> {
+        let mut res = 0;
+        res += get_mut_arcmutex!(self.draft).activate_adapters(adapters.clone())?;
+        res += get_mut_arcmutex!(self.target).activate_adapters(adapters)?;
+        Ok(res)
+    }
+}
+
+impl MetadataMixin for SpeculativePipeline {
+    fn device(&self) -> Device {
+        get_mut_arcmutex!(self.target).device()
+    }
+    fn tokenizer(&self) -> Arc<Tokenizer> {
+        get_mut_arcmutex!(self.target).tokenizer()
+    }
+    fn name(&self) -> String {
+        format!(
+            "Speculative: tgt = `{}`, draft = `{}`, gamma = `{}`",
+            get_mut_arcmutex!(self.target).name(),
+            get_mut_arcmutex!(self.draft).name(),
+            self.gamma,
+        )
+    }
+    fn reset_non_granular_state(&self) {
+        get_mut_arcmutex!(self.target).reset_non_granular_state();
+        get_mut_arcmutex!(self.draft).reset_non_granular_state();
+    }
+    fn get_metadata(&self) -> &GeneralMetadata {
+        &self.metadata
+    }
+}
+
 #[async_trait::async_trait]
 impl Pipeline for SpeculativePipeline {
+    fn forward_inputs(&mut self, _inputs: Box<dyn Any>) -> Result<Tensor> {
+        unreachable!()
+    }
+    async fn sample(
+        &self,
+        _seqs: &mut [&mut Sequence],
+        _logits: Tensor,
+        _prefix_cacher: &mut PrefixCacheManager,
+        _disable_eos_stop: bool,
+        _rng: Arc<std::sync::Mutex<Isaac64Rng>>,
+    ) -> Result<()> {
+        unreachable!()
+    }
     async fn step(
         &mut self,
         input_seqs: &mut [&mut Sequence],
@@ -233,16 +328,18 @@ impl Pipeline for SpeculativePipeline {
             let is_xlora = get_mut_arcmutex!(self.draft).get_metadata().is_xlora;
             let device = get_mut_arcmutex!(self.draft).device();
             let has_no_kv_cache = get_mut_arcmutex!(self.draft).get_metadata().has_no_kv_cache;
-            let inputs = calculate_inputs(
-                &[seq],
-                is_prompt && i == 0, // Only prompt (no kv cache) if first
-                is_xlora,
-                &device,
-                has_no_kv_cache,
-                None,
-            )
-            .unwrap();
-            let logits = get_mut_arcmutex!(self.draft).forward_inputs(inputs)?;
+            let inputs = self
+                .get_input_processor()
+                .process_inputs(
+                    &[seq],
+                    is_prompt && i == 0, // Only prompt (no kv cache) if first
+                    is_xlora,
+                    &device,
+                    has_no_kv_cache,
+                    None,
+                )
+                .unwrap();
+            let logits = get_mut_arcmutex!(self.draft).forward_inputs(Box::new(inputs))?;
 
             let sample = sample_sequence(
                 logits.clone(),
@@ -294,17 +391,19 @@ impl Pipeline for SpeculativePipeline {
         let has_no_kv_cache = get_mut_arcmutex!(self.target)
             .get_metadata()
             .has_no_kv_cache;
-        let inputs = calculate_inputs(
-            &[seq],
-            true, // use the "prefill" tokens
-            is_xlora,
-            &device,
-            has_no_kv_cache,
-            Some((self.gamma, initial_cache_len)), // Get the last gamma, see above
-        )
-        .unwrap();
+        let inputs = self
+            .get_input_processor()
+            .process_inputs(
+                &[seq],
+                true, // use the "prefill" tokens
+                is_xlora,
+                &device,
+                has_no_kv_cache,
+                Some((self.gamma, initial_cache_len)), // Get the last gamma, see above
+            )
+            .unwrap();
 
-        let logits = get_mut_arcmutex!(self.target).forward_inputs(inputs)?;
+        let logits = get_mut_arcmutex!(self.target).forward_inputs(Box::new(inputs))?;
 
         // Reset the prefill tokens
         seq.reset_prefill_toks();
@@ -449,80 +548,5 @@ impl Pipeline for SpeculativePipeline {
         // - Maybe fixed up cache of base model based on accepted tokens.
 
         Ok(())
-    }
-    fn forward_inputs(&mut self, _: ModelInputs) -> anyhow::Result<Tensor, candle_core::Error> {
-        unreachable!("Speculative pipeline handles the forward pass in `.step`")
-    }
-    async fn sample(
-        &self,
-        _seqs: &mut [&mut Sequence],
-        _logits: Tensor,
-        _prefix_cacher: &mut PrefixCacheManager,
-        _disable_eos_stop: bool,
-        __rng: Arc<Mutex<Isaac64Rng>>,
-    ) -> Result<()> {
-        unreachable!("Speculative pipeline handles sampling in `.step`")
-    }
-    fn device(&self) -> Device {
-        get_mut_arcmutex!(self.target).device()
-    }
-    fn tokenizer(&self) -> Arc<Tokenizer> {
-        get_mut_arcmutex!(self.target).tokenizer()
-    }
-    fn name(&self) -> String {
-        format!(
-            "Speculative: tgt = `{}`, draft = `{}`, gamma = `{}`",
-            get_mut_arcmutex!(self.target).name(),
-            get_mut_arcmutex!(self.draft).name(),
-            self.gamma,
-        )
-    }
-    fn get_chat_template(&self) -> Arc<ChatTemplate> {
-        get_mut_arcmutex!(self.target).get_chat_template()
-    }
-    fn reset_non_granular_state(&self) {
-        get_mut_arcmutex!(self.target).reset_non_granular_state();
-        get_mut_arcmutex!(self.draft).reset_non_granular_state();
-    }
-    fn re_isq_model(&mut self, dtype: GgmlDType) -> anyhow::Result<()> {
-        get_mut_arcmutex!(self.target).re_isq_model(dtype)?;
-        get_mut_arcmutex!(self.draft).re_isq_model(dtype)
-    }
-    fn get_metadata(&self) -> &GeneralMetadata {
-        &self.metadata
-    }
-    fn clone_in_cache(&mut self, seqs: &mut [&mut Sequence], modify_draft_cache: bool) {
-        DefaultCacheManager.clone_in_cache(
-            &mut *get_mut_arcmutex!(self.draft),
-            seqs,
-            modify_draft_cache,
-        );
-        DefaultCacheManager.clone_in_cache(&mut *get_mut_arcmutex!(self.target), seqs, false);
-    }
-    fn clone_out_cache(&mut self, seqs: &mut [&mut Sequence], modify_draft_cache: bool) {
-        DefaultCacheManager.clone_out_cache(
-            &mut *get_mut_arcmutex!(self.draft),
-            seqs,
-            modify_draft_cache,
-        );
-        DefaultCacheManager.clone_out_cache(&mut *get_mut_arcmutex!(self.target), seqs, false);
-    }
-    fn set_none_cache(&mut self, reset_non_granular: bool, modify_draft_cache: bool) {
-        DefaultCacheManager.set_none_cache(&mut *get_mut_arcmutex!(self.draft), modify_draft_cache);
-        DefaultCacheManager.set_none_cache(&mut *get_mut_arcmutex!(self.target), false);
-        if reset_non_granular {
-            self.reset_non_granular_state()
-        }
-        self.latest_logit_cache = None;
-    }
-    fn cache(&self) -> &Cache {
-        unreachable!()
-    }
-    /// Returns the number of activated adapters.
-    fn activate_adapters(&mut self, adapters: Vec<String>) -> anyhow::Result<usize> {
-        let mut res = 0;
-        res += get_mut_arcmutex!(self.draft).activate_adapters(adapters.clone())?;
-        res += get_mut_arcmutex!(self.target).activate_adapters(adapters)?;
-        Ok(res)
     }
 }
