+++ mistralrs-core/src/xlora_models/mixtral.rs
@@ -3,6 +3,7 @@
 use crate::{
     layers::{MatMul, ScaledDotProductAttention},
     lora::{linear_no_bias, LinearLayerLike, LoraConfig, Ordering},
+    pipeline::IsqModel,
 };
 /// Mixtral Model
 /// https://github.com/huggingface/transformers/blob/main/src/transformers/models/mixtral/modeling_mixtral.py
@@ -803,6 +804,43 @@ impl XLoraModel {
     }
 }
 
+impl IsqModel for XLoraModel {
+    fn get_tensors(&mut self) -> (Vec<(&mut QMatMul, Option<usize>)>, &dyn DeviceMapper) {
+        let mut tensors = Vec::new();
+        tensors.push((&mut self.lm_head, None));
+        for (i, layer) in self.layers.iter_mut().enumerate() {
+            tensors.push((
+                Arc::get_mut(&mut layer.self_attn.q_proj).unwrap().inner(),
+                Some(i),
+            ));
+            tensors.push((
+                Arc::get_mut(&mut layer.self_attn.k_proj).unwrap().inner(),
+                Some(i),
+            ));
+            tensors.push((
+                Arc::get_mut(&mut layer.self_attn.v_proj).unwrap().inner(),
+                Some(i),
+            ));
+            tensors.push((
+                Arc::get_mut(&mut layer.self_attn.o_proj).unwrap().inner(),
+                Some(i),
+            ));
+            tensors.push((
+                Arc::get_mut(&mut layer.block_sparse_moe.gate)
+                    .unwrap()
+                    .inner(),
+                Some(i),
+            ));
+            for expert in &mut layer.block_sparse_moe.experts {
+                tensors.push((Arc::get_mut(&mut expert.w1).unwrap().inner(), Some(i)));
+                tensors.push((Arc::get_mut(&mut expert.w2).unwrap().inner(), Some(i)));
+                tensors.push((Arc::get_mut(&mut expert.w3).unwrap().inner(), Some(i)));
+            }
+        }
+        (tensors, &*self.mapper)
+    }
+}
+
 impl NormalModel for XLoraModel {
     fn forward(
         &mut self,
@@ -850,40 +888,6 @@ impl NormalModel for XLoraModel {
     }
     fn max_seq_len(&self) -> usize {
         self.max_seq_len
-    }
-    fn get_tensors(&mut self) -> (Vec<(&mut QMatMul, Option<usize>)>, &dyn DeviceMapper) {
-        let mut tensors = Vec::new();
-        tensors.push((&mut self.lm_head, None));
-        for (i, layer) in self.layers.iter_mut().enumerate() {
-            tensors.push((
-                Arc::get_mut(&mut layer.self_attn.q_proj).unwrap().inner(),
-                Some(i),
-            ));
-            tensors.push((
-                Arc::get_mut(&mut layer.self_attn.k_proj).unwrap().inner(),
-                Some(i),
-            ));
-            tensors.push((
-                Arc::get_mut(&mut layer.self_attn.v_proj).unwrap().inner(),
-                Some(i),
-            ));
-            tensors.push((
-                Arc::get_mut(&mut layer.self_attn.o_proj).unwrap().inner(),
-                Some(i),
-            ));
-            tensors.push((
-                Arc::get_mut(&mut layer.block_sparse_moe.gate)
-                    .unwrap()
-                    .inner(),
-                Some(i),
-            ));
-            for expert in &mut layer.block_sparse_moe.experts {
-                tensors.push((Arc::get_mut(&mut expert.w1).unwrap().inner(), Some(i)));
-                tensors.push((Arc::get_mut(&mut expert.w2).unwrap().inner(), Some(i)));
-                tensors.push((Arc::get_mut(&mut expert.w3).unwrap().inner(), Some(i)));
-            }
-        }
-        (tensors, &*self.mapper)
     }
     fn activate_adapters(&mut self, adapter_names: Vec<String>) -> Result<usize> {
         let mut sum = 0;
